feat(switchMapTo): add higher-order lettable version of switchMapTo
