+++ src/operators/index.ts
@@ -64,9 +64,11 @@ export { skip } from './skip';
 export { skipLast } from './skipLast';
 export { skipUntil } from './skipUntil';
 export { skipWhile } from './skipWhile';
+export { startWith } from './startWith';
 export { subscribeOn } from './subscribeOn';
 export { switchAll } from './switchAll';
 export { switchMap } from './switchMap';
+export { switchMapTo } from './switchMapTo';
 export { takeLast } from './takeLast';
 export { tap } from './tap';
 export { timestamp } from './timestamp';
